Package the script
