+++ arista_vtep_update/__main__.py
@@ -1,0 +1,5 @@
+import sys
+from . import main
+
+if __name__ == "__main__":
+    sys.exit(main(sys.argv[1:]))
